덱리스트 페이지로 납치 당하는 문제 수정

+++ lib/page/main_page.dart
@@ -58,7 +58,7 @@ class _MainPageState extends State<MainPage> {
       routes: [DeckBuilderRoute(deck: null), DeckListRoute(), CollectRoute(), InfoRoute()],
       builder: (context, child, controller) {
         controller.addListener(() {
-          if (controller.indexIsChanging) {
+          if (controller.indexIsChanging && mounted) {
             _cardOverlayService.removeAllOverlays();
           }
         });
